+++ CHANGELOG.md
@@ -6,11 +6,31 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Add Theme Provider + Theme Switcher on Header for development (should be in settings in future)
+### Changed
+
+### Removed
+
+## [0.10.4.8] - 2025-10-04
+
+### Added
+
+- Add Theme Provider + Theme Switcher on Header for development (should be in settings in future).
+
+- Add Chatbot UI with copy message action.
+
+- Implement Conversation features: Rename, Archive, Create new Conversation.
+
+- Implement ChatTriz features: Chat.
+
+- Add audio recorder with visualizer and transcript.
+
+- Add warning chatbot content: “Nội dung AI cung cấp chỉ mang tính chất tham khảo. Có thể cung cấp một số câu trả lời sai lệch”.
 
 ### Changed
 
 ### Removed
+
+- Remove package "immer".
 
 ## [0.9.22.19] - 2025-09-22